requirements.txt
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
prometheus_client==0.21.1
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.38

